[feature] add command checker on command handler
