+++ SSD/SSD/SSD.vcxproj
@@ -131,6 +131,7 @@
     <ClCompile Include="cmd_executor.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">  %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
+    <ClCompile Include="cmd_checker.cpp" />
     <ClCompile Include="command_handler.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
