+++ SSD/SSD/cmd_executor.cpp
@@ -4,58 +4,6 @@
 #include "output_handler.cpp"
 #include "nand_handler.cpp"
 
-class CmdChecker {
-public:
-	static bool isValidLBA(int lba) {
-		return isValidRange(lba);
-	}
-
-	static bool isValidValue(string value) {
-		if (false == isHexaFormat(value)) {
-			return false;
-		}
-		else if (10 != value.length()) {
-			return false;
-		}
-		else if (false == isHexaData(value.substr(2))) {
-			return false;
-		}
-
-		return true;
-	}
-
-private:
-	static bool isValidRange(int lba) {
-		if (0 > lba || 100 <= lba) {
-			return false;
-		}
-
-		return true;
-	}
-
-	static bool isHexaFormat(string value) {
-		if (0 != value.find("0x")) {
-			return false;
-		}
-
-		return true;
-	}
-
-	static bool isHexaData(string value) {
-		for (int i = 0; i < value.length(); i++) {
-			if (('0' <= value[i] && '9' >= value[i]) ||
-				('a' <= value[i] && 'f' >= value[i]) ||
-				('A' <= value[i] && 'F' >= value[i])) {
-				continue;
-			}
-
-			return false;
-		}
-
-		return true;
-	}
-};
-
 class CmdExecutor {
 public:
 	CmdExecutor() {
@@ -64,10 +12,6 @@ public:
 	}
 
 	virtual string read(int lba) {
-		if (false == CmdChecker::isValidLBA(lba)) {
-			throw std::out_of_range("[READ ERROR] Out of lba");
-		}
-
 		string ret = readDataOnAddr(lba);
 
 		m_outputHandler->write(ret);
@@ -76,13 +20,6 @@ public:
 	}
 
 	virtual void write(int lba, string value) {
-		if (false == CmdChecker::isValidLBA(lba)) {
-			throw std::out_of_range("[WRITE ERROR] Out of lba");
-		}
-		if (false == CmdChecker::isValidValue(value)) {
-			throw std::invalid_argument("[WRITE ERROR] Invalid value to write.");
-		}
-
 		writeDataOnAddr(lba, value);
 	}
